services/chat-service/modules/speech.py
"""
语音服务模块 - 阿里云ASR/TTS
从speech-service迁移
"""
import os
import base64
from typing import Optional, List
import structlog

# 阿里云NLS SDK
try:
    from aliyun.speech.transcriber import Transcriber
    from aliyun.speech.synthesizer import Synthesizer
except ImportError:
    Transcriber = None
    Synthesizer = None

logger = structlog.get_logger()

# 阿里云配置
ALIYUN_ACCESS_KEY_ID = os.getenv("ALIYUN_ACCESS_KEY_ID", "")
ALIYUN_ACCESS_KEY_SECRET = os.getenv("ALIYUN_ACCESS_KEY_SECRET", "")
ASR_APP_KEY = os.getenv("ALIYUN_ASR_APP_KEY", "")
TTS_APP_KEY = os.getenv("ALIYUN_TTS_APP_KEY", "")

# 音色列表
VOICE_LIST = [
    {"voice_id": "zhixiaobai", "name": "知小白", "gender": "female", "language": "zh-CN", "description": "清新女声"},
    {"voice_id": "zhixiaoxia", "name": "知小夏", "gender": "female", "language": "zh-CN", "description": "温柔女声"},
    {"voice_id": "zhimiao", "name": "知妙", "gender": "female", "language": "zh-CN", "description": "甜美童声"},
    {"voice_id": "yunge", "name": "云哥", "gender": "male", "language": "zh-CN", "description": "沉稳男声"},
    {"voice_id": "yunjie", "name": "云杰", "gender": "male", "language": "zh-CN", "description": "磁性男声"},
    {"voice_id": "aisha", "name": "Aisha", "gender": "female", "language": "en-US", "description": "英文女声"},
]


async def transcribe(audio_data: str, sample_rate: int = 16000, format: str = "wav") -> str:
    """
    语音识别 - 阿里云ASR
    
    Args:
        audio_data: Base64编码的WAV音频数据
        sample_rate: 采样率，默认16000
        format: 音频格式，默认wav
    
    Returns:
        识别文本
    """
    if not audio_data:
        raise ValueError("未提供音频数据")
    
    try:
        # 解码Base64音频
        audio_bytes = base64.b64decode(audio_data)
        
        # 如果有SDK，使用真实ASR
        if Transcriber:
            transcriber = Transcriber(
                appkey=ASR_APP_KEY,
                token=get_aliyun_token(),
                sample_rate=sample_rate
            )
            result = transcriber.transcribe(audio_bytes)
            text = result.get("text", "")
        else:
            # 模拟实现（用于测试）
            text = f"[模拟识别] 音频长度: {len(audio_bytes)} 字节, 采样率: {sample_rate}Hz"
        
        logger.info("transcribe_success", text_length=len(text))
        return text
        
    except Exception as e:
        logger.error("transcribe_error", error=str(e))
        raise Exception(f"识别失败: {str(e)}")


async def synthesize(text: str, voice_id: str = "zhixiaobai", sample_rate: int = 16000, format: str = "wav") -> str:
    """
    语音合成 - 阿里云TTS
    
    Args:
        text: 要合成的文本
        voice_id: 音色ID，默认zhixiaobai
        sample_rate: 采样率，默认16000
        format: 音频格式，默认wav
    
    Returns:
        Base64编码的音频数据
    """
    if not text:
        raise ValueError("未提供文本")
    
    try:
        # 如果有SDK，使用真实TTS
        if Synthesizer:
            synthesizer = Synthesizer(
                appkey=TTS_APP_KEY,
                token=get_aliyun_token(),
                voice=voice_id,
                sample_rate=sample_rate
            )
            audio_data = synthesizer.synthesize(text)
            # 返回Base64编码的音频
            audio_base64 = base64.b64encode(audio_data).decode('utf-8')
        else:
            # 模拟实现（用于测试）
            audio_base64 = base64.b64encode(f"[模拟音频] {text}".encode()).decode('utf-8')
        
        logger.info("synthesize_success", voice_id=voice_id, text_length=len(text))
        return audio_base64
        
    except Exception as e:
        logger.error("synthesize_error", error=str(e))
        raise Exception(f"合成失败: {str(e)}")


async def list_voices() -> List[dict]:
    """
    获取可用音色列表
    
    Returns:
        音色列表
    """
    return VOICE_LIST


def get_aliyun_token():
    """
    获取阿里云Token（简化实现，实际应该使用STS）
    
    Returns:
        Token字符串
    """
    # 实际项目中应该调用STS服务获取临时Token
    return ""
