services/chat-service/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Temporal
temporalio==1.4.0

# MongoDB
motor==3.3.2
pymongo==4.6.1

# 阿里云语音服务
aliyun-nls-python-sdk==1.0.0

# LangChain + 通义千问 (指定版本)
langchain==0.3.20
langchain-core==0.3.44
langchain-openai==0.3.8
langchain-google-genai==2.1.0

# Milvus
pymilvus==2.3.5

# Mem0
mem0ai==0.0.9

# HTTP客户端
httpx==0.26.0

# 日志
structlog==24.1.0

# 工具类
python-dotenv==1.0.0
python-multipart==0.0.6
